- Make `TLVRecord` templated and adjust all layers that depend on it - Convert PPPoE tags to use the TLV infra

+++ Packet++/src/TLVData.cpp
@@ -11,35 +11,35 @@ TLVRecordBuilder::TLVRecordBuilder()
 	m_RecValue = NULL;
 }
 
-TLVRecordBuilder::TLVRecordBuilder(uint8_t recType, const uint8_t* recValue, uint8_t recValueLen)
+TLVRecordBuilder::TLVRecordBuilder(uint32_t recType, const uint8_t* recValue, uint8_t recValueLen)
 {
 	init(recType, recValue, recValueLen);
 }
 
-TLVRecordBuilder::TLVRecordBuilder(uint8_t recType, uint8_t recValue)
+TLVRecordBuilder::TLVRecordBuilder(uint32_t recType, uint8_t recValue)
 {
 	init(recType, &recValue, sizeof(uint8_t));
 }
 
-TLVRecordBuilder::TLVRecordBuilder(uint8_t recType, uint16_t recValue)
+TLVRecordBuilder::TLVRecordBuilder(uint32_t recType, uint16_t recValue)
 {
 	recValue = htobe16(recValue);
 	init(recType, (uint8_t*)&recValue, sizeof(uint16_t));
 }
 
-TLVRecordBuilder::TLVRecordBuilder(uint8_t recType, uint32_t recValue)
+TLVRecordBuilder::TLVRecordBuilder(uint32_t recType, uint32_t recValue)
 {
 	recValue = htobe32(recValue);
 	init(recType, (uint8_t*)&recValue, sizeof(uint32_t));
 }
 
-TLVRecordBuilder::TLVRecordBuilder(uint8_t recType, const IPv4Address& recValue)
+TLVRecordBuilder::TLVRecordBuilder(uint32_t recType, const IPv4Address& recValue)
 {
 	uint32_t recIntValue = recValue.toInt();
 	init(recType, (uint8_t*)&recIntValue, sizeof(uint32_t));
 }
 
-TLVRecordBuilder::TLVRecordBuilder(uint8_t recType, const std::string& recValue)
+TLVRecordBuilder::TLVRecordBuilder(uint32_t recType, const std::string& recValue)
 {
 	uint8_t* recValueByteArr = (uint8_t*)recValue.c_str();
 	init(recType, recValueByteArr, recValue.length());
@@ -80,7 +80,7 @@ TLVRecordBuilder::~TLVRecordBuilder()
 	if (m_RecValue != NULL) delete [] m_RecValue;
 }
 
-void TLVRecordBuilder::init(uint8_t recType, const uint8_t* recValue, uint8_t recValueLen)
+void TLVRecordBuilder::init(uint32_t recType, const uint8_t* recValue, size_t recValueLen)
 {
 	m_RecType = recType;
 	m_RecValueLen = recValueLen;
